+++ src/main/java/chess/WebUIChessApplication.java
@@ -11,8 +11,8 @@ import chess.domain.Board;
 import chess.domain.Team;
 import chess.domain.Turn;
 import chess.domain.position.Position;
+import chess.domain.state.BoardRepository;
 import chess.dto.UnitDto;
-import chess.state.BoardRepository;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import spark.ModelAndView;
