+++ src/main/java/chess/domain/state/ChessGameState.java
@@ -1,4 +1,4 @@
-package chess.state;
+package chess.domain.state;
 
 import java.util.List;
