+++ src/main/java/chess/view/InputView.java
@@ -4,9 +4,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 
+import chess.domain.state.Command;
 import chess.dto.RequestDto;
 import chess.exception.IllegalCommandException;
-import chess.state.Command;
 
 public class InputView {
 	private static final Scanner scanner = new Scanner(System.in);
