+++ src/test/java/chess/domain/state/BoardRepositoryTest.java
@@ -1,4 +1,4 @@
-package chess.state;
+package chess.domain.state;
 
 import static org.assertj.core.api.Assertions.*;
